refactor itemlist page into .tsx
